+++ Back_End/src/main/java/com/ssafy/youniverse/dto/res/ActorResDto.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class ActorResDto {
-    private Integer actorId;
+    private Long actorId;
     private String actorName;
     private String actorImage;
 }
