+++ Back_End/src/main/java/com/ssafy/youniverse/dto/res/FollowResDto.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class FollowResDto {
-    private Integer followId;
+    private Long followId;
     private MemberSimpleResDto followerResDto;
     private MemberSimpleResDto followingResDto;
 }
