+++ Back_End/src/main/java/com/ssafy/youniverse/dto/res/HeartMovieResDto.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class HeartMovieResDto {
-    private Integer heartMovieId;
+    private Long heartMovieId;
     private MemberSimpleResDto memberSimpleResDto;
     private MovieSimpleResDto movieSimpleResDto;
 }
