+++ Back_End/src/main/java/com/ssafy/youniverse/repository/DirectorMovieRepository.java
@@ -3,5 +3,5 @@ package com.ssafy.youniverse.repository;
 import com.ssafy.youniverse.entity.DirectorMovie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DirectorMovieRepository extends JpaRepository<DirectorMovie, Integer> {
+public interface DirectorMovieRepository extends JpaRepository<DirectorMovie, Long> {
 }
