+++ Back_End/src/main/java/com/ssafy/youniverse/repository/KeywordMemberRepository.java
@@ -3,5 +3,5 @@ package com.ssafy.youniverse.repository;
 import com.ssafy.youniverse.entity.KeywordMember;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface KeywordMemberRepository extends JpaRepository<KeywordMember, Integer> {
+public interface KeywordMemberRepository extends JpaRepository<KeywordMember, Long> {
 }
