+++ Back_End/src/main/java/com/ssafy/youniverse/repository/MemberRepositoryCustom.java
@@ -6,5 +6,5 @@ import org.springframework.data.domain.Pageable;
 
 public interface MemberRepositoryCustom {
     Page<Member> findAllQueryDsl(Pageable pageable, String keyword, String nickname, String total);
-    Integer findByRandomQueryDsl();
+    Long findByRandomQueryDsl();
 }
