+++ Back_End/src/main/java/com/ssafy/youniverse/repository/YoutubeKeywordRepository.java
@@ -3,5 +3,5 @@ package com.ssafy.youniverse.repository;
 import com.ssafy.youniverse.entity.YoutubeKeyword;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface YoutubeKeywordRepository extends JpaRepository<YoutubeKeyword, Integer> {
+public interface YoutubeKeywordRepository extends JpaRepository<YoutubeKeyword, Long> {
 }
